+++ NetWork/ExAFNetworking/CacheProvider.m
@@ -206,7 +206,7 @@ CacheProvider * _sharedCacheProvider = nil;
 {
 	const char* str = [input UTF8String];
     unsigned char result[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(str, strlen(str), result);
+    CC_MD5(str, (int)strlen(str), result);
     NSMutableString *ret = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH];
 
     for(int i = 0; i<CC_MD5_DIGEST_LENGTH; i++) {
